Add extra check
The size == 1 check just below is required for scripted to work fine.
(The first module report must correspond to the passed dependency for
scripted to find the sbt launcher JAR.)

It seems under some conditions, the sbt launcher dependency ends up
being duplicated (it's already duplicated in the allDependencies task,
that should be set by sbt). This in turn makes the check fail, and makes
scripted not find the sbt launcher. This happens with
playframework/interplay at commit d782b75.

The call to distinct circumvents that case.