Add explicit dependency for checkBuildSources
Rather than adding some custom logic to EvaluateTask.runTask, it makes
more sense to explicitly add a dependency on checkBuildSources to every
(non-injected) task in the build. This both reduces the amount of custom
logic for the checkBuildSources task and also makes it show up in the
`inspect` command

I also reordered the injected settings so that they come before the
project settings. This should allow the user to override the injected
settings if they wanted. The injected settings do not have the
dependency on checkBuildSources added which allows the other file stamp
tasks to run concurrently with the build source check. This means that
for tasks like compile there is often effectively zero overhead for the
build source check.

Finally, I simplified Continuous so that there is no custom logic for
tracking the meta build. By adding the dependency on checkBuildSources
to every task, the tasks end up automatically watching the build files
anyway.